requirements.txt
azure-functions==1.8.0
requests==2.19.1